+++ lib/changelog/parsers/intridea_format.rb
@@ -17,7 +17,10 @@ module Danger
 
             changelog_line = ChangelogLineParser.parse(line)
 
-            if changelog_line.nil? || changelog_line.invalid?
+            if changelog_line.nil?
+              notify_of_bad_line line, 'cannot be parsed'
+              next
+            elsif changelog_line.invalid?
               detail = changelog_line.validation_result.to_s if changelog_line.validation_result&.invalid?
               notify_of_bad_line line, detail
               next